requirements.txt
azure-core==1.17.0
azure-functions==1.7.2
azure-storage-blob==12.8.1
azure-identity==1.7.1
azure-keyvault-secrets==4.3.0
certifi==2021.5.30
cffi
charset-normalizer==2.0.4
cryptography==3.4.7
idna==3.2
isodate==0.6.0
msrest==0.6.21
oauthlib==3.1.1
pandas==1.5.3
pycparser==2.20
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
urllib3==1.26.6
pyodbc==5.0.0
SQLAlchemy==1.4.29
